+++ src/algorithms/find_closest_with_paths.hpp
@@ -32,7 +32,7 @@ using namespace std;
     /// Search does not exceed max_search_dist bases.
     ///
     /// Will only ever return an empty vector or a 1-element vector.
-    vector<tuple<handle_t, int64_t, bool>> find_closest_with_paths(const PathHandleGraph& graph, handle_t start, size_t offset, size_t max_search_dist);
+    vector<tuple<handle_t, int64_t, bool>> find_closest_with_paths(const PathHandleGraph& graph, handle_t start, size_t offset, int64_t max_search_dist);
     
 
 }
